README and image support
Created first half of README corresponding to first half of
workshop. This includes a table of contents for participants to easily
follow along with, instructions to install and set up an android studio
project, and a description of the content that wil be covered in the
workshop slides.

Activities, Layouts, and Views described in detail and supporting images
provided for participants.